Group the workspace files by pane
