+++ front/app/containers/Admin/projects/project/backofficeRedesign/ProjectWorkspace/Header/ShareDropdown/SharePanel.tsx
@@ -20,9 +20,9 @@ import useFeatureFlag from 'hooks/useFeatureFlag';
 import { trackEventByName } from 'utils/analytics';
 import { useIntl } from 'utils/cl-intl';
 
-import projectHeaderMessages from '../../../projectHeader/messages';
-import shareLinkTracks from '../../../projectHeader/ShareLink/tracks';
-import messages from '../messages';
+import projectHeaderMessages from '../../../../projectHeader/messages';
+import shareLinkTracks from '../../../../projectHeader/ShareLink/tracks';
+import messages from '../../messages';
 
 import AccessPeopleRow from './AccessPeopleRow';
